test(plugins): gate the SIGTERM-delivery test to unix builds
Branch review flagged that the test skipped only Windows while assuming a
POSIX shell and signal delivery; it now lives behind the same //go:build
unix tag as the terminateGit implementation it pins.
